+++ src/services/game.service.ts
@@ -1,6 +1,9 @@
 import { inject, injectable } from "tsyringe";
-import GameRepository from "@/repositories/game.repository";
 import { Game } from "@/entities/game.entity";
+import GameRepository from "@/repositories/game.repository";
+import GameDomain from "@/domains/game.domain";
+import ConflictException from "@/exceptions/conflict.exception";
+import ValidationException from "@/exceptions/validation.exception";
 
 @injectable()
 export default class GameService {
@@ -9,6 +12,18 @@ export default class GameService {
   ) {}
 
   async createGame(data: Partial<Game>): Promise<Game> {
+    const existing = await this.gameRepository.findByName(data.name!);
+
+    if (existing) {
+      throw new ConflictException("Game name already registered");
+    }
+
+    if (!GameDomain.doesHaveValidName(data.name!)) {
+      throw new ValidationException(
+        "Name cannot contain numbers, capitalized and special characters (except for underscores)",
+      );
+    }
+
     return this.gameRepository.create(data);
   }
 
@@ -21,10 +36,28 @@ export default class GameService {
   }
 
   async updateGame(id: number, data: Partial<Game>): Promise<Game | null> {
-    return this.gameRepository.update(id, data);
+    const existing = await this.gameRepository.findById(id);
+
+    if (!existing) {
+      throw new ConflictException("Game not found");
+    }
+
+    if (data.name && !GameDomain.doesHaveValidName(data.name)) {
+      throw new ValidationException(
+        "Name cannot contain numbers, capitalized and special characters (except for underscores)",
+      );
+    }
+
+    return this.gameRepository.update(id, { ...existing, ...data });
   }
 
   async deleteGame(id: number): Promise<boolean> {
+    const existing = await this.gameRepository.findById(id);
+
+    if (!existing) {
+      throw new ConflictException("Game not found");
+    }
+
     return this.gameRepository.delete(id);
   }
 }
